reformat random macid/cid functions
