+++ pycid/random/__init__.py
@@ -1,3 +1,5 @@
-from pycid.random.random_cid import random_cid, random_cids, random_macid, random_macidbase, random_macids  # noqa
+from pycid.random.random_cid import random_cid, random_cids  # noqa
 from pycid.random.random_cpd import RandomCPD  # noqa
 from pycid.random.random_dag import random_dag  # noqa
+from pycid.random.random_macid import random_macid, random_macids  # noqa
+from pycid.random.random_macidbase import random_macidbase  # noqa
